Funzionalità agazziniere completate

+++ src/main/java/remoteInterfaces/UserServiceRemote.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserServiceRemote {
     void addUser(Utente utente);
-    Utente findUserById(int id);
+    Utente findUserById(Long id);
     Utente findUserByEmail(String email);
     void updateUser(Utente utente);
     void removeUser(Utente utente);
